--- api/alembic/versions/0c254e3da7f2_13_unique_email_deleted.py
@@ -1,32 +1,0 @@
-"""13_unique_email_deleted
-
-Revision ID: 0c254e3da7f2
-Revises: f0b4c7b8c283
-Create Date: 2024-05-03 14:04:42.961693
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '0c254e3da7f2'
-down_revision: Union[str, None] = 'f0b4c7b8c283'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('usuario_usuario_email_key', 'usuario_usuario', type_='unique')
-    op.create_unique_constraint('_email_deleted_uc.', 'usuario_usuario', ['email', 'deleted'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('_email_deleted_uc.', 'usuario_usuario', type_='unique')
-    op.create_unique_constraint('usuario_usuario_email_key', 'usuario_usuario', ['email'])
-    # ### end Alembic commands ###
