--- api/alembic/versions/16ad060d293c_10_add_campo_valido_documento.py
@@ -1,30 +1,0 @@
-"""10_add_campo_valido_documento
-
-Revision ID: 16ad060d293c
-Revises: 53801e5c9a81
-Create Date: 2024-05-03 17:56:39.001105
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '16ad060d293c'
-down_revision: Union[str, None] = '53801e5c9a81'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mercado_mercado', sa.Column('mercado_valido', sa.Boolean(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mercado_mercado', 'mercado_valido')
-    # ### end Alembic commands ###
