--- api/alembic/versions/183465e7c359_mercado_big_integer.py
@@ -1,36 +1,0 @@
-"""mercado_big_integer
-
-Revision ID: 183465e7c359
-Revises: aea8113813f8
-Create Date: 2024-04-19 15:11:51.789450
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '183465e7c359'
-down_revision: Union[str, None] = 'aea8113813f8'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('mercado_mercado', 'telefone',
-               existing_type=sa.INTEGER(),
-               type_=sa.BigInteger(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('mercado_mercado', 'telefone',
-               existing_type=sa.BigInteger(),
-               type_=sa.INTEGER(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
