--- api/alembic/versions/29604cd89eb7_mercado_perfil.py
@@ -1,32 +1,0 @@
-"""mercado_perfil
-
-Revision ID: 29604cd89eb7
-Revises: 183465e7c359
-Create Date: 2024-04-19 15:44:16.101222
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '29604cd89eb7'
-down_revision: Union[str, None] = '183465e7c359'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mercado_mercado', sa.Column('descricao', sa.String(length=500), nullable=True))
-    op.add_column('mercado_mercado', sa.Column('logo', sa.String(length=255), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mercado_mercado', 'logo')
-    op.drop_column('mercado_mercado', 'descricao')
-    # ### end Alembic commands ###
