--- api/alembic/versions/2efa2c087256_13_add_campo.py
@@ -1,30 +1,0 @@
-"""13_add_campo
-
-Revision ID: 2efa2c087256
-Revises: c2faa0e3d910
-Create Date: 2024-05-06 10:02:06.303919
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2efa2c087256'
-down_revision: Union[str, None] = 'c2faa0e3d910'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mercado_produto_promocao_erp', sa.Column('descricao', sa.String(length=500), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mercado_produto_promocao_erp', 'descricao')
-    # ### end Alembic commands ###
