--- api/alembic/versions/53801e5c9a81_9_produto_consumidor.py
@@ -1,79 +1,0 @@
-"""9_produto_consumidor
-
-Revision ID: 53801e5c9a81
-Revises: ca26d80e0aca
-Create Date: 2024-04-23 13:12:52.562594
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import fastapi_storages
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-
-# revision identifiers, used by Alembic.
-revision: str = '53801e5c9a81'
-down_revision: Union[str, None] = 'c01e4be735cd'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-storage = fastapi_storages.FileSystemStorage(path="./media")
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('mercado_promocao',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('produto', postgresql.JSONB(astext_type=sa.Text()), nullable=False),
-    sa.Column('data_inicial', sa.DateTime(), nullable=False),
-    sa.Column('data_final', sa.DateTime(), nullable=False),
-    sa.Column('percentual_desconto', sa.Float(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.Column('deleted', sa.Boolean(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('mercado_produto',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('mercado_id', sa.UUID(), nullable=True),
-    sa.Column('nome', sa.String(length=255), nullable=True),
-    sa.Column('marca', sa.String(length=255), nullable=True),
-    sa.Column('data_validade', sa.DateTime(), nullable=True),
-    sa.Column('ncm_produto', sa.String(length=10), nullable=True),
-    sa.Column('gtin_produto', sa.String(length=14), nullable=True),
-    sa.Column('mpn_produto', sa.String(length=30), nullable=True),
-    sa.Column('id_produto_erp', sa.String(), nullable=True),
-    sa.Column('descricao', sa.String(length=500), nullable=True),
-    sa.Column('preco', sa.Float(), nullable=True),
-    sa.Column('preco_promocional', sa.Float(), nullable=True),
-    sa.Column('imagem', fastapi_storages.integrations.sqlalchemy.ImageType(storage=storage), nullable=True),
-    sa.Column('codigo_produto', sa.String(length=30), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.Column('deleted', sa.Boolean(), nullable=False),
-    sa.ForeignKeyConstraint(['mercado_id'], ['mercado_mercado.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('usuario_consumidor',
-    sa.Column('id', sa.UUID(), autoincrement=False, nullable=False),
-    sa.Column('cep', sa.VARCHAR(length=8), autoincrement=False, nullable=False),
-    sa.Column('estado', sa.VARCHAR(length=255), autoincrement=False, nullable=False),
-    sa.Column('cidade', sa.VARCHAR(length=255), autoincrement=False, nullable=False),
-    sa.Column('bairro', sa.VARCHAR(length=255), autoincrement=False, nullable=False),
-    sa.Column('endereco', sa.VARCHAR(length=255), autoincrement=False, nullable=False),
-    sa.Column('numero_endereco', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('complemento', sa.VARCHAR(length=255), autoincrement=False, nullable=True),
-    sa.Column('telefone', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('id_usuario', sa.UUID(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['id_usuario'], ['usuario_usuario.id'], name='usuario_consumidor_id_usuario_fkey'),
-    sa.PrimaryKeyConstraint('id', name='usuario_consumidor_pkey')
-    )
-    op.drop_table('mercado_produto')
-    op.drop_table('mercado_promocao')
-    # ### end Alembic commands ###
