--- api/alembic/versions/9bd9b04018b2_fix_8_consumidor.py
@@ -1,36 +1,0 @@
-"""fix_8_consumidor
-
-Revision ID: 9bd9b04018b2
-Revises: e1109dc11d9f
-Create Date: 2024-04-22 12:07:01.734298
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '9bd9b04018b2'
-down_revision: Union[str, None] = 'e1109dc11d9f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('usuario_consumidor', sa.Column('id_usuario', sa.UUID(), nullable=False))
-    op.drop_constraint('usuario_consumidor_id_fkey', 'usuario_consumidor', type_='foreignkey')
-    op.create_foreign_key(None, 'usuario_consumidor', 'usuario_usuario', ['id_usuario'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'usuario_consumidor', type_='foreignkey')
-    op.create_foreign_key('usuario_consumidor_id_fkey', 'usuario_consumidor', 'usuario_usuario', ['id'], ['id'])
-    op.drop_column('usuario_consumidor', 'id_usuario')
-    # ### end Alembic commands ###
-    
-
