--- api/alembic/versions/aea8113813f8_second_migration.py
@@ -1,69 +1,0 @@
-"""second_migration
-
-Revision ID: aea8113813f8
-Revises: c822c0d21489
-Create Date: 2024-04-12 18:48:01.982324
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'aea8113813f8'
-down_revision: Union[str, None] = 'c822c0d21489'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('mercado_mercado',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('usuario_id', sa.UUID(), nullable=True),
-    sa.Column('cnpj', sa.String(length=14), nullable=False),
-    sa.Column('razao_social', sa.String(length=255), nullable=False),
-    sa.Column('nome_fantasia', sa.String(length=255), nullable=False),
-    sa.Column('telefone', sa.Integer(), nullable=False),
-    sa.Column('cep', sa.String(length=8), nullable=False),
-    sa.Column('estado', sa.String(length=255), nullable=False),
-    sa.Column('cidade', sa.String(length=255), nullable=False),
-    sa.Column('bairro', sa.String(length=255), nullable=False),
-    sa.Column('endereco', sa.String(length=255), nullable=False),
-    sa.Column('numero_endereco', sa.Integer(), nullable=False),
-    sa.Column('complemento', sa.String(length=255), nullable=True),
-    sa.Column('nome_responsavel', sa.String(length=255), nullable=False),
-    sa.Column('cpf_responsavel', sa.String(length=11), nullable=False),
-    sa.ForeignKeyConstraint(['usuario_id'], ['usuario_usuario.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('cnpj')
-    )
-    op.create_table('usuario_consumidor',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('cep', sa.String(length=8), nullable=False),
-    sa.Column('estado', sa.String(length=255), nullable=False),
-    sa.Column('cidade', sa.String(length=255), nullable=False),
-    sa.Column('bairro', sa.String(length=255), nullable=False),
-    sa.Column('endereco', sa.String(length=255), nullable=False),
-    sa.Column('numero_endereco', sa.Integer(), nullable=False),
-    sa.Column('complemento', sa.String(length=255), nullable=True),
-    sa.Column('telefone', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['id'], ['usuario_usuario.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.add_column('usuario_usuario', sa.Column('dono_mercado', sa.Boolean(), nullable=False))
-    op.drop_constraint('unique_email_deleted', 'usuario_usuario', type_='unique')
-    op.create_unique_constraint(None, 'usuario_usuario', ['email'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'usuario_usuario', type_='unique')
-    op.create_unique_constraint('unique_email_deleted', 'usuario_usuario', ['email', 'deleted'])
-    op.drop_column('usuario_usuario', 'dono_mercado')
-    op.drop_table('usuario_consumidor')
-    op.drop_table('mercado_mercado')
-    # ### end Alembic commands ###
