--- api/alembic/versions/b6c270e1e861_14_fix_usuario_email_deleted_constraint.py
@@ -1,32 +1,0 @@
-"""14_fix_usuario_email_deleted_constraint
-
-Revision ID: b6c270e1e861
-Revises: 0c254e3da7f2
-Create Date: 2024-05-03 17:02:22.434160
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b6c270e1e861'
-down_revision: Union[str, None] = '0c254e3da7f2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('_email_deleted_uc.', 'usuario_usuario', type_='unique')
-    op.create_index('_email_deleted_unique', 'usuario_usuario', ['email', 'deleted'], unique=True, postgresql_where=sa.text('NOT deleted'))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('_email_deleted_unique', table_name='usuario_usuario', postgresql_where=sa.text('NOT deleted'))
-    op.create_unique_constraint('_email_deleted_uc.', 'usuario_usuario', ['email', 'deleted'])
-    # ### end Alembic commands ###
