--- api/alembic/versions/b8d5c5b4f7c9_third_migration_model_google.py
@@ -1,36 +1,0 @@
-"""third_migration_model_google
-
-Revision ID: b8d5c5b4f7c9
-Revises: aea8113813f8
-Create Date: 2024-04-19 13:58:09.274250
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b8d5c5b4f7c9'
-down_revision: Union[str, None] = 'aea8113813f8'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('usuario_auth_google',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('id_google', sa.String(length=255), nullable=False),
-    sa.Column('id_usuario', sa.UUID(), nullable=False),
-    sa.ForeignKeyConstraint(['id_usuario'], ['usuario_usuario.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('usuario_auth_google')
-    # ### end Alembic commands ###
