--- api/alembic/versions/c01e4be735cd_fix_9_consumidor_telefone.py
@@ -1,36 +1,0 @@
-"""fix_9_consumidor_telefone
-
-Revision ID: c01e4be735cd
-Revises: 9bd9b04018b2
-Create Date: 2024-04-22 12:13:48.756401
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c01e4be735cd'
-down_revision: Union[str, None] = '9bd9b04018b2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('usuario_consumidor', 'telefone',
-               existing_type=sa.INTEGER(),
-               type_=sa.BigInteger(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('usuario_consumidor', 'telefone',
-               existing_type=sa.BigInteger(),
-               type_=sa.INTEGER(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
