--- api/alembic/versions/c2faa0e3d910_12_model_produtos_er.py
@@ -1,53 +1,0 @@
-"""12_model_produtos_er
-
-Revision ID: c2faa0e3d910
-Revises: 16ad060d293c
-Create Date: 2024-05-06 08:53:01.885650
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c2faa0e3d910'
-down_revision: Union[str, None] = '16ad060d293c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('mercado_produto_promocao_erp',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('mercado_id', sa.UUID(), nullable=True),
-    sa.Column('nome', sa.String(length=255), nullable=True),
-    sa.Column('preco', sa.Float(), nullable=True),
-    sa.Column('preco_promocional', sa.Float(), nullable=True),
-    sa.Column('codigo_produto', sa.String(length=30), nullable=True),
-    sa.Column('ncm_produto', sa.String(length=10), nullable=True),
-    sa.Column('id_produto_erp', sa.String(), nullable=True),
-    sa.Column('marca', sa.String(length=255), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.Column('deleted', sa.Boolean(), nullable=False),
-    sa.ForeignKeyConstraint(['mercado_id'], ['mercado_mercado.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.alter_column('usuario_consumidor', 'telefone',
-               existing_type=sa.INTEGER(),
-               type_=sa.BigInteger(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('usuario_consumidor', 'telefone',
-               existing_type=sa.BigInteger(),
-               type_=sa.INTEGER(),
-               existing_nullable=False)
-    op.drop_table('mercado_produto_promocao_erp')
-    # ### end Alembic commands ###
