--- api/alembic/versions/c52d04666796_mercado_documentos.py
@@ -1,52 +1,0 @@
-"""mercado_documentos
-
-Revision ID: c52d04666796
-Revises: ede676064b89
-Create Date: 2024-04-19 22:47:45.869246
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c52d04666796'
-down_revision: Union[str, None] = 'ede676064b89'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('mercado_mercado', 'cnpj',
-               existing_type=sa.VARCHAR(length=18),
-               type_=sa.String(length=14),
-               existing_nullable=False)
-    op.alter_column('mercado_mercado', 'razao_social',
-               existing_type=sa.VARCHAR(length=255),
-               type_=sa.String(length=30),
-               existing_nullable=False)
-    op.alter_column('mercado_mercado', 'nome_fantasia',
-               existing_type=sa.VARCHAR(length=255),
-               type_=sa.String(length=55),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('mercado_mercado', 'nome_fantasia',
-               existing_type=sa.String(length=55),
-               type_=sa.VARCHAR(length=255),
-               existing_nullable=False)
-    op.alter_column('mercado_mercado', 'razao_social',
-               existing_type=sa.String(length=30),
-               type_=sa.VARCHAR(length=255),
-               existing_nullable=False)
-    op.alter_column('mercado_mercado', 'cnpj',
-               existing_type=sa.String(length=14),
-               type_=sa.VARCHAR(length=18),
-               existing_nullable=False)
-    # ### end Alembic commands ###
