--- api/alembic/versions/ede676064b89_mercado_created_on.py
@@ -1,42 +1,0 @@
-"""mercado_created_on
-
-Revision ID: ede676064b89
-Revises: 29604cd89eb7
-Create Date: 2024-04-19 22:23:29.235134
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'ede676064b89'
-down_revision: Union[str, None] = '29604cd89eb7'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mercado_mercado', sa.Column('created_at', sa.DateTime(), nullable=False))
-    op.add_column('mercado_mercado', sa.Column('updated_at', sa.DateTime(), nullable=False))
-    op.add_column('mercado_mercado', sa.Column('deleted', sa.Boolean(), nullable=False))
-    op.alter_column('mercado_mercado', 'cnpj',
-               existing_type=sa.VARCHAR(length=14),
-               type_=sa.String(length=18),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('mercado_mercado', 'cnpj',
-               existing_type=sa.String(length=18),
-               type_=sa.VARCHAR(length=14),
-               existing_nullable=False)
-    op.drop_column('mercado_mercado', 'deleted')
-    op.drop_column('mercado_mercado', 'updated_at')
-    op.drop_column('mercado_mercado', 'created_at')
-    # ### end Alembic commands ###
