--- api/alembic/versions/f0b4c7b8c283_fix_12_pk_consumidor.py
@@ -1,34 +1,0 @@
-"""fix_12_pk_consumidor
-
-Revision ID: f0b4c7b8c283
-Revises: ae0abbc1bdda
-Create Date: 2024-05-03 11:35:46.891766
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f0b4c7b8c283'
-down_revision: Union[str, None] = 'ae0abbc1bdda'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('usuario_consumidor_id_usuario_fkey', 'usuario_consumidor', type_='foreignkey')
-    op.create_foreign_key(None, 'usuario_consumidor', 'usuario_usuario', ['id'], ['id'])
-    op.drop_column('usuario_consumidor', 'id_usuario')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('usuario_consumidor', sa.Column('id_usuario', sa.UUID(), autoincrement=False, nullable=False))
-    op.drop_constraint(None, 'usuario_consumidor', type_='foreignkey')
-    op.create_foreign_key('usuario_consumidor_id_usuario_fkey', 'usuario_consumidor', 'usuario_usuario', ['id_usuario'], ['id'])
-    # ### end Alembic commands ###
