+++ app/ExpoGo/app.json
@@ -47,7 +47,7 @@
       "eas": {
         "projectId": "be0ffba9-faf2-48d2-848c-75f6a0841938"
       }
-    }
-    
+    },
+    "owner": "coutinho.cassio"
   }
 }
