+++ app/ExpoGo/package.json
